Move business logic into a library
Moves the openVulnQuery tool's business logic into a library, so as to
avoid implicit relative imports that break Python 3 compatibility.

This move preserves existing script and package behavior.
openVulnQuery/openVulnQuery/__init__.py was modified so that the
`openVulnQuery` namespace would remain as it currently is when
`openVulnQuery` is imported as a package.

The command `python openVulnQuery/openVulnQuery/main.py` also still
executes the `main(...)` function of `openVulnQuery.main`, and if the
packages is installed using `setuptools`, the `openVulnQuery` command
will have its current behavior.
