+++ openVulnQuery/openVulnQuery/_library/constants.py
