+++ openVulnQuery/openVulnQuery/_library/query_client.py
@@ -1,16 +1,16 @@
+import datetime as dt
+import json
 import logging
+import os
+import uuid
 
 import requests
 
-import advisory
-import authorization
-import config
-import constants
-import rest_api
-import json
-import os
-import uuid
-import datetime as dt
+from . import advisory
+from . import authorization
+from . import config
+from . import constants
+from . import rest_api
 
 ADV_TOKENS = constants.ADVISORY_FORMAT_TOKENS
