+++ openVulnQuery/openVulnQuery/_library/utils.py
@@ -3,7 +3,7 @@ import json
 import sys
 from platform import python_version_tuple as p_v_t
 
-import constants
+from . import constants
 
 __is_future_version = False if int(p_v_t()[0]) < 3 else True
